refactor(liblili2json): move dictionary related function to new file
