+++ workspaces/liblili2json/liblili2json/__init__.py
@@ -7,6 +7,7 @@ import itertools
 from slugify import slugify
 from .types import *
 from .utils import *
+from .dictionary import *
 
 __version__ = '0.1.0'
 
@@ -254,27 +255,3 @@ def get_skill_data(skill, course):
             len(skill.words), len(skill.phrases)),
         "challenges": get_challenges_data(skill, course),
     }
-
-
-def define_words_in_sentence(course, sentence, reverse):
-    return [define_word(course, word, reverse) for word in sentence.split()]
-
-
-def get_dictionary_item(course, word, reverse):
-    dictionary_item = list(
-        filter(
-            lambda item: item.word == word and item.reverse == reverse,
-            course.dictionary))
-
-    return dictionary_item[0] if dictionary_item else None
-
-
-def define_word(course, word, reverse):
-    dictionary_item = get_dictionary_item(course, word, reverse)
-    if dictionary_item and dictionary_item.definition:
-        return {
-            "word": word,
-            "definition": dictionary_item.definition
-        }
-
-    return {"word": word}
